fix: add log level to the monodeploy configuration
