+++ packages/cli/src/cli.test.ts
@@ -71,6 +71,7 @@ describe('CLI', () => {
                     "tag": false,
                   },
                   "jobs": 6,
+                  "logLevel": "0",
                   "maxConcurrentReads": 3,
                   "maxConcurrentWrites": 4,
                   "packageGroupManifestField": "group",
@@ -120,6 +121,7 @@ describe('CLI', () => {
                     "tag": undefined,
                   },
                   "jobs": 0,
+                  "logLevel": undefined,
                   "maxConcurrentReads": 0,
                   "maxConcurrentWrites": 0,
                   "packageGroupManifestField": undefined,
@@ -276,6 +278,7 @@ describe('CLI', () => {
                         "tag": true,
                       },
                       "jobs": 6,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 3,
                       "maxConcurrentWrites": 5,
                       "packageGroupManifestField": undefined,
@@ -373,6 +376,7 @@ describe('CLI', () => {
                         "tag": false,
                       },
                       "jobs": 6,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 6,
                       "maxConcurrentWrites": 2,
                       "packageGroupManifestField": undefined,
@@ -457,6 +461,7 @@ describe('CLI', () => {
                         "tag": false,
                       },
                       "jobs": 6,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 2,
                       "maxConcurrentWrites": 1,
                       "packageGroupManifestField": "group",
@@ -546,6 +551,7 @@ describe('CLI', () => {
                         "tag": true,
                       },
                       "jobs": 6,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 3,
                       "maxConcurrentWrites": 5,
                       "packageGroupManifestField": undefined,
@@ -637,6 +643,7 @@ describe('CLI', () => {
                         "tag": false,
                       },
                       "jobs": 3,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 10,
                       "maxConcurrentWrites": 11,
                       "packageGroupManifestField": "group",
@@ -724,6 +731,7 @@ describe('CLI', () => {
                         "tag": false,
                       },
                       "jobs": 3,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 10,
                       "maxConcurrentWrites": 11,
                       "packageGroupManifestField": undefined,
@@ -890,6 +898,7 @@ describe('CLI', () => {
                         "tag": false,
                       },
                       "jobs": 2,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 0,
                       "maxConcurrentWrites": 1,
                       "packageGroupManifestField": "group",
@@ -990,6 +999,7 @@ describe('CLI', () => {
                         "tag": false,
                       },
                       "jobs": 2,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 0,
                       "maxConcurrentWrites": 1,
                       "packageGroupManifestField": "group",
@@ -1077,6 +1087,7 @@ describe('CLI', () => {
                         "tag": true,
                       },
                       "jobs": 6,
+                      "logLevel": undefined,
                       "maxConcurrentReads": 3,
                       "maxConcurrentWrites": 5,
                       "packageGroupManifestField": undefined,
